requirements.txt
numpy==1.26.4
Pillow==10.3.0
matplotlib==3.8.4
scikit-learn==1.4.2
opencv-python-headless==4.9.0.80
streamlit==1.33.0
mtcnn==0.1.1
tensorflow==2.16.1
